apps/blegps: Refactor data handling
This is the first step to make GPS data avaiable for both OLED and
Bluetooth (LN Service).

The data is tentatively parsed using minmea in an interrupt and then
it is decided whether enough data is retrieved to make an update in
main task or not. If so, the main task will call an update function
to convert data from nimnea to properly scaled numbers etc. to be
used by application. These data are available globally so both OLED
routine and BLE routines can access the same data without need to
parse them again.