apps/blegps: Simplify update code
The GPS data were updated after some NMEA messages so they were
actually updated few times during the same sequence. Now we just have
single timer which updates data (and OLED+BLE) once a second, no
matter what is received from GPS.

In this case we may actually update GPS in the middle of NMEA sequence
but this is not a big problem - we can improve this later.